+++ ios/Fletcher/Fletcher/UI/SettingsView.swift
@@ -50,7 +50,7 @@ struct SettingsView: View {
                 Section {
                     HStack {
                         Spacer()
-                        Text("Fletcher v1.2.3")
+                        Text("Fletcher v1.2.4")
                             .font(.footnote)
                             .foregroundColor(.secondary)
                         Spacer()